# lib/api-zod/src/generated/types/notifyPrefs.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HomardClaw private agent office API
 * OpenAPI spec version: 0.1.0
 */

export interface NotifyPrefs {
  onCompleted: boolean;
  onFailed: boolean;
  onBlocked: boolean;
  onApprovalNeeded: boolean;
}