# lib/api-zod/src/generated/types/schedule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HomardClaw private agent office API
 * OpenAPI spec version: 0.1.0
 */
import type { NotifyPrefs } from './notifyPrefs';
import type { ScheduleCadence } from './scheduleCadence';

export interface Schedule {
  id: string;
  name: string;
  agentId: string;
  agentName: string;
  objective: string;
  priority: string;
  /** @nullable */
  providerOverride?: string | null;
  /** @nullable */
  modelOverride?: string | null;
  /** @nullable */
  budgetCents?: number | null;
  cadence: ScheduleCadence;
  timezone: string;
  /** @nullable */
  runAt?: Date | null;
  /** @nullable */
  timeOfDay?: string | null;
  /** @nullable */
  daysOfWeek?: number[] | null;
  /** @nullable */
  dayOfMonth?: number | null;
  notify: NotifyPrefs;
  enabled: boolean;
  /** @nullable */
  nextRunAt?: Date | null;
  /** @nullable */
  lastRunAt?: Date | null;
  /** @nullable */
  lastTaskId?: string | null;
  /** @nullable */
  lastTaskStatus?: string | null;
  createdAt: Date;
}