# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HomardClaw private agent office API
 * OpenAPI spec version: 0.1.0
 */
import type { NotifyPrefs } from './notifyPrefs';
import type { ScheduleInputCadence } from './scheduleInputCadence';
import type { ScheduleInputPriority } from './scheduleInputPriority';
import type { ScheduleInputProviderOverride } from './scheduleInputProviderOverride';

export interface ScheduleInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  agentId: string;
  /**
     * @minLength 3
     * @maxLength 5000
     */
  objective: string;
  priority?: ScheduleInputPriority;
  providerOverride?: ScheduleInputProviderOverride;
  /** @maxLength 200 */
  modelOverride?: string;
  /**
     * @minimum 0.01
     * @maximum 1000000
     */
  budgetCents?: number;
  cadence: ScheduleInputCadence;
  /**
     * @minLength 1
     * @maxLength 60
     */
  timezone: string;
  runAt?: Date;
  /** @pattern ^([01]?\d|2[0-3]):[0-5]\d$ */
  timeOfDay?: string;
  /**
     * @items.minimum 0
     * @items.maximum 6
     */
  daysOfWeek?: number[];
  /**
     * @minimum 1
     * @maximum 31
     */
  dayOfMonth?: number;
  notify?: NotifyPrefs;
}