# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HomardClaw private agent office API
 * OpenAPI spec version: 0.1.0
 */
import type { NotifyPrefs } from './notifyPrefs';
import type { ScheduleUpdateCadence } from './scheduleUpdateCadence';
import type { ScheduleUpdatePriority } from './scheduleUpdatePriority';

export interface ScheduleUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /**
     * @minLength 3
     * @maxLength 5000
     */
  objective?: string;
  priority?: ScheduleUpdatePriority;
  /** @nullable */
  providerOverride?: string | null;
  /** @nullable */
  modelOverride?: string | null;
  /** @nullable */
  budgetCents?: number | null;
  cadence?: ScheduleUpdateCadence;
  /**
     * @minLength 1
     * @maxLength 60
     */
  timezone?: string;
  runAt?: Date;
  /** @pattern ^([01]?\d|2[0-3]):[0-5]\d$ */
  timeOfDay?: string;
  /**
     * @items.minimum 0
     * @items.maximum 6
     */
  daysOfWeek?: number[];
  /**
     * @minimum 1
     * @maximum 31
     */
  dayOfMonth?: number;
  notify?: NotifyPrefs;
  enabled?: boolean;
}